Fork pie-spring-2023 and delete fluff
Fork the repo for the spring competition.
Delete peripheral code since our new robot will
not have an arm or hand.
Delete input code because it's tragically messy
and this is a good chance to rewrite it.
Delete chassis code because we will use an angled
omnidirectional design for the new robot.
Delete path code because that was overcomplicated
and broken, and there's no way we're getting that
fixed in eight days.
Delete main.py. We're going to start mostly from
scratch and our software architecture will look
dramatically different.
Delete qual_finals_code.py. Being preprocessor
output, it should never have been committed in
the first place.